Change field label and placeholder in AutoComplete
Updated the label and placeholder for the AutoComplete field in 'examples/basic/src/main.rs'. Previously, the field was for entering tags, but now it is dedicated to entering country code. The change also contains an updated help string pointing out that country code data is sourced from 'https://restcountries.com/v3.1/name/'.